Introduce an EnvironmentStack
Environments (scopes) need to be able to defer to enclosing environments during symbol lookup. The guide
suggests implementing this using a pointer-to-parent, so a singly linked list. I'm going to try something
different and pass around an explicit stack of environments and have the stack be the thing you query. Rust's
docs suggest that you typically want a Vec or a VecDeque, and indeed I'm not reordering the linked list so
there's not really any benefit.

Maybe splitting hairs for a hobby project, but still.