A bit of a rethink
Move EVAL into evaluator.
Yes, different steps eval in different ways. But I don't think that's interesting enough to merit passing the
Context around. (Also it was easier that fighting the borrow checker. It could be that I'm missing something.

Ditch EnvironmentStack in favour of an Environment as outlined in the guide.
The existence of closures means we'll need to keep multiple references around to the environment in which the
closure was defined. It's just not the case that there's always a single, linear hierarchy of environments.
I'm not sure if I got the lifetime annotations right here... let's see.

Should have been a separate change, but oh well.