Step 5: tail-call optimisation
That was easier than I was expecting. I'm not sure if it was right to use `Cow` in the EVAL loop, and I'm
still concerned that I'm cloning things all over the place. So I'm glad that we're wrapping containers with an
Rc. (Though it's still a toy language and project so this is maybe a bit of a moot point.)