Use tuple structs rather than type aliases
If I understand rightly, this combined with an `impl Deref` lets you treat e.g. a `struct Pixels(u64)` as a
u64, but not the other way round. I get the sense that the boilerplate to do so is much smaller than in C++.
Might be misunderstanding though.

I made this in the middle of a lot of other changes so I don't think this commit will compile perfectly.