Have (if x y x) fully EVAL y and z
I don't fully understand why this fixes things, because the bug I was seeing involved EVALling too far. Hmmm.
But it's a clearly a bug: I'd expect (if x (prn "truthy") (prn "falsy")) to print something!

It's a shame mal doesn't have a test case to cover this!

Looks like it regressed in f50a530e when I was trying to be clever with TCO. Beware, Icarus!